backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
pymysql==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
aiofiles==24.1.0
pypdf==5.1.0

